+++ CHANGELOG.md
@@ -10,23 +10,30 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Added
 
-- BC-155 add filter for task where the teacher is not a substitution teacher
+- BC-155 - add filter for task where the teacher is not a substitution teacher
 
 ### Changed
 
-### Fixed
+- BC-501 - displaying topic name in task overview
 
 ## [26.13.0] - 2021-10-19
 
 ### Added
 
 - BC-428 - bug-adjust-tint-for International
+- BC-495 - ungrouping dashboard elements for room overview
 
 ### Changed
 
 ### Fixed
 
 - BC-393 - cleaned up some translations
+
+## [26.12.1] - 2021-10-12
+
+### Changed
+
+- BC-379 - hard-code node 14 and npm 6
 
 ## [26.12.0] - 2021-10-12